Added the dashboard page and route.Modified login and registration html and route. Added the database to models.py
