+++ project/templates/login.html
@@ -16,7 +16,19 @@
                 <input type="password" name="password" placeholder="Password" class="mb-4 p-2 w-full border rounded">
                 <button type="submit" name="login" class="bg-blue-500 text-white py-2 px-4 rounded w-full">Login</button>
             </form>
+            <p class="mt-4"></p>
+            {% with messages = get_flashed_messages() %}
+              {% if messages %}
+                <ul class="flashes">
+                  {% for message in messages %}
+                   <li>{{ message }}</li>
+              {% endfor %}
+           </ul>
+       {% endif %}
+     {% endwith %}
+
             <p class="mt-4">Don't have an account? <a href="{{ url_for('register') }}" class="text-blue-600">Register here</a></p>
+            <p class="mt-4">Go Back to Home Page - <a href="{{ url_for('index') }}" class="text-blue-600">Home Page</a></p>
         </div>
     </div>
 </body>
